made the changes to move fireCannon function to another file - clean up
